Repacking stage 2. Added listener package and more - read below.
Huge changes here... And even greater changes are comming.

Project now consisnts of packages of standart classes
for theit respective jobs. Each individual object SHOULD
be able to exitst as a unit, not requiring hard links to others.

This is made possible due to usage of listeners,
where, for example Interface implements PaintListener,
and a root game Container implements and UpdateListener.

Now concerning containres:
1). Non-position based (ghosts):
	Entity
	Container extends Entity
2). Position based (real objects):
	Object
	Layer extends Object

This needs to be figured out, but basicly an object is painted
relative to it's Layer (might be renamed to Box),
and Layers (being Objects), also render relative to their owners...

This is a huge rework... Everything is still working correctly,
but everything might fail shortly.

BUGFIX: Keyboard input not working on Linux fixed:
	KeyboardListener now listens for the whole JFrame,
	rather than JPanel. Other listeners left intact.
