+++ src/com/valentine/game/GameWorld.java
@@ -4,9 +4,11 @@ import java.awt.Dimension;
 import java.awt.event.KeyEvent;
 import java.awt.event.MouseEvent;
 
-import com.valentine.game.gameworld.*;
+import com.valentine.game.entity.Entity;
+import com.valentine.game.listener.InputListener;
+import com.valentine.game.world.*;
 
-public abstract class GameWorld implements GameInputListener, Entity {
+public abstract class GameWorld implements InputListener, Entity {
 	
 	protected static Dimension dimension;
 	protected static boolean ready = false;
