+++ src/com/valentine/game/entity/Entity.java
@@ -1,0 +1,5 @@
+package com.valentine.game.entity;
+
+import com.valentine.game.listener.*;
+
+public interface Entity extends PaintListener, UpdateListener {}
