--- src/com/valentine/game/gameworld/Entity.java
@@ -1,9 +1,0 @@
-package com.valentine.game.gameworld;
-
-import java.awt.*;
-
-public interface Entity {
-	public void update();
-	public void paint(Graphics _graphics);
-	
-}
