--- src/com/valentine/game/gameworld/GameInputListener.java
@@ -1,9 +1,0 @@
-package com.valentine.game.gameworld;
-
-import java.awt.event.KeyListener;
-import java.awt.event.MouseListener;
-import java.awt.event.MouseMotionListener;
-
-public interface GameInputListener extends KeyListener, MouseListener, MouseMotionListener {
-	
-}
