+++ src/com/valentine/game/listener/InputListener.java
@@ -1,0 +1,9 @@
+package com.valentine.game.listener;
+
+import java.awt.event.KeyListener;
+import java.awt.event.MouseListener;
+import java.awt.event.MouseMotionListener;
+
+public interface InputListener extends KeyListener, MouseListener, MouseMotionListener {
+	
+}
