+++ src/com/valentine/game/listener/PaintListener.java
@@ -1,0 +1,8 @@
+package com.valentine.game.listener;
+
+import java.awt.Graphics;
+
+public interface PaintListener
+{
+	public void paint(Graphics _graphics);
+}
