+++ src/com/valentine/game/listener/UpdateListener.java
@@ -1,0 +1,6 @@
+package com.valentine.game.listener;
+
+public interface UpdateListener
+{
+	public void update();
+}
