+++ src/com/valentine/game/world/GameWorldImpl_BallsAndWorlds.java
@@ -1,4 +1,4 @@
-package com.valentine.game.gameworld;
+package com.valentine.game.world;
 
 import java.awt.Graphics;
 import java.awt.event.KeyEvent;
@@ -6,9 +6,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 
 import com.valentine.game.GameWorld;
-import com.valentine.game.gameworld.entity.Circle;
-import com.valentine.game.gameworld.entity.HelloWorld;
-import com.valentine.game.gameworld.entity.Player;
+import com.valentine.game.entity.Circle;
+import com.valentine.game.entity.Entity;
+import com.valentine.game.entity.HelloWorld;
+import com.valentine.game.entity.Player;
 
 public class GameWorldImpl_BallsAndWorlds extends GameWorld {
